re-construct file structure to use multiple AI features
