+++ data_discovery_ai/common/keyword_classification_parameters.ini
@@ -12,8 +12,8 @@ fl_gamma = 2
 fl_alpha = 0.7
 epoch = 100
 batch = 32
-early_stopping_patience = 5
+early_stopping_patience = 3
 reduce_lr_patience = 5
 validation_split = 0.2
-confidence = 0.4
-top_N = 2
+confidence = 0.5
+top_N = 1
